All events are handled (including pocket cards and common cards), until rounds change.

+++ trunk/cspoker-server/src/org/cspoker/server/game/playerCommunication/PlayerCommunication.java
@@ -174,6 +174,6 @@ public interface PlayerCommunication {
 	 * @return
 	 * @throws IllegalActionException
 	 */
-	public List<GameEvent> getLatestGameEventsAndAck(long ack) throws IllegalActionException;
+	public List<GameEvent> getLatestGameEventsAndAck(int ack) throws IllegalActionException;
 
 }
